+++ js/utils.js
@@ -1,0 +1,9 @@
+
+//функция для рандомного значения
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+export {getRandomInteger};
